+++ arduino/asserv/asservCommon/PID.h
@@ -16,18 +16,18 @@ bound : limite la valeur de l'intégrale entre +max_I et -max_I
 
 class PID{
 	public:
-	void setPID(double n_P, double n_I, double n_D);
-	void setBias(double n_bias);
+	void setPID(float n_P, float n_I, float n_D);
+	void setBias(float n_bias);
 	void reset();
-	double getOutput();
-	double compute(double error);
-	PID(double n_p = 0, double n_I = 0, double n_D = 0, double n_bias = 0);
+	float getOutput();
+	float compute(float error);
+	PID(float n_p = 0, float n_I = 0, float n_D = 0, float n_bias = 0);
 	private:
-	double P, I, D;
-	double last_error; //derniere erreur pour la (D)ériée
-	double output;
-	double error_I; //somme des erreurs * intervale = (I)ntégrale
-	double bias; //somme constante ajoutée au résultat du PID
+	float P, I, D;
+	float last_error; //derniere erreur pour la (D)ériée
+	float output;
+	float error_I; //somme des erreurs * intervale = (I)ntégrale
+	float bias; //somme constante ajoutée au résultat du PID
 	bool initDone; //permet d'éviter les erreurs dues à l'absence de dérivée au premier compute
 };
