+++ arduino/asserv/asservCommon/local_math.cpp
@@ -7,7 +7,7 @@
 #include "local_math.h"
 #include <math.h>
 
-double moduloTwoPI(double angle){
+float moduloTwoPI(float angle){
 	if(angle >= 0)
 		while(angle >= M_PI)
 			angle -= 2.0*M_PI;
