+++ arduino/asserv/asservCommon/robotstate.h
@@ -10,22 +10,22 @@
 #include "encoder.h"
 #include <math.h>
 
-const double ENC_TICKS_TO_MM = (2.0*M_PI*(ENC_RADIUS))/(TICKS_PER_TURN);// = mm/ticks
-const double ENC_MM_TO_TICKS = (TICKS_PER_TURN)/(2.0*M_PI*(ENC_RADIUS)); // = ticks/mm
+const float ENC_TICKS_TO_MM = (2.0*M_PI*(ENC_RADIUS))/(TICKS_PER_TURN);// = mm/ticks
+const float ENC_MM_TO_TICKS = (TICKS_PER_TURN)/(2.0*M_PI*(ENC_RADIUS)); // = ticks/mm
 
 typedef struct pos pos;
 struct pos{
 	long x;
 	long y;
-	double angle;
-	double rangle;
+	float angle;
+	float rangle;
 };
 
 typedef struct m_pos m_pos;
 struct m_pos{
-	double x;
-	double y;
-	double angle;
+	float x;
+	float y;
+	float angle;
 };
 
 class RobotState{
